refactor: split ruby-to-blocks-converter/index.js into 6 files
Extract converter-errors.js (messages, constants), comment-handler.js
(comment extraction/association), class-visitor.js (visitClassNode),
initialize-handler.js (def initialize processing), and module-visitor.js
(visitModuleNode) from the 1580-line index.js.

Refs #357
